// src/shared/api/generated/model/createRoomRequestDto.ts
/**
 * Generated by orval 🍺
 * Do not edit manually.
 * Yoclick API
 * API multi-centro de reservas en marca blanca. Contrato canónico para la app.
 * OpenAPI spec version: 0.1.0
 */

export interface CreateRoomRequestDto {
  /**
     * @minLength 1
     * @maxLength 80
     */
  name: string;
  /**
     * @minimum 1
     * @maximum 500
     */
  capacity?: number;
}